pass all floats throught bigdecimal

+++ src/main/java/com/synrc/bert/Enc.java
@@ -16,7 +16,7 @@ public interface Enc<T> extends F<T, Term> {
 
     public static Enc<String>   stringEnc = Term::str;
     public static Enc<byte[]>   binEnc  = Term::bin;
-    public static Enc<Double>   floatEnc = Term::float754;
+    public static Enc<BigDecimal>   floatEnc = Term::float754;
     public static Enc<BigDecimal>   floatStrEnc = Term::floatStr;
     public static Enc<Byte>     byteEnc = Term::bt;
     public static Enc<Integer>  intEnc = Term::in;
